完善32位、64位下release和debug版本工程配置

+++ src/Sheeps/ClientProtocolSub.cpp
@@ -328,7 +328,7 @@ bool CreateFileDirectory(const char* dir)
 	if ((nPathLen = strlen(path)) < 1)
 		return false;
 
-	for (int i = 0; i < nPathLen; ++i)
+	for (size_t i = 0; i < nPathLen; ++i)
 	{
 		if (path[i] == '\\')
 		{
